+++ src/main/java/dk/mfaester/grav/Gnyf.java
@@ -25,8 +25,8 @@ public class Gnyf extends AbstractDrawable {
     }
 
     @Override
-    protected byte[] createIndices() {
-        final byte[] indices = {
+    protected int[] createIndices() {
+        final int[] indices = {
                 0, 1, 2,
         };
         return indices;
